@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Subscriber;
 use Illuminate\Database\Seeder;
 
@@ -7,9 +9,9 @@ class LeaderboardSeeder extends Seeder
 {
     public function run()
     {
-        factory(Subscriber::class)->create();
+        Subscriber::factory()->create();
         foreach (range(1, 49) as $i) {
-            factory(Subscriber::class)->create([
+            Subscriber::factory()->create([
                 'referred_by' => Subscriber::all()->random()->referrer_id,
                 'pledged' => true,
                 'hide_from_pledge_board' => false,
