@@ -1,17 +1,26 @@
 <?php
 
-use Faker\Generator as Faker;
-use App\Subscriber;
+namespace Database\Factories;
 
-$factory->define(App\Message::class, function (Faker $faker) {
-    return [
-        'twilio_sid'        => $faker->md5(),
-        'subscriber_number' => function () {
-            return factory(Subscriber::class)->create()->number;
-        },
-        'incoming' => false,
-        'to'       => $faker->phoneNumber(),
-        'from'     => $faker->phoneNumber(),
-        'body'     => $faker->sentence(),
-    ];
-});
+use App\Message;
+use App\Subscriber;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class MessageFactory extends Factory
+{
+    protected $model = Message::class;
+
+    public function definition()
+    {
+        return [
+            'twilio_sid' => $this->faker->md5(),
+            'subscriber_number' => function () {
+                return Subscriber::factory()->create()->number;
+            },
+            'incoming' => false,
+            'to' => $this->faker->phoneNumber(),
+            'from' => $this->faker->phoneNumber(),
+            'body' => $this->faker->sentence(),
+        ];
+    }
+}
