@@ -1,7 +1,9 @@
 <?php
 
-use Illuminate\Database\Seeder;
+namespace Database\Seeders;
+
 use App\Tag;
+use Illuminate\Database\Seeder;
 
 class TagsTableSeeder extends Seeder
 {
