@@ -1,8 +1,10 @@
 <?php
 
+namespace Database\Seeders;
+
+use App\User;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\Hash;
-use App\User;
 
 class UsersTableSeeder extends Seeder
 {
