@@ -2,37 +2,18 @@
 
 namespace Tests;
 
+use App\User;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
 
 abstract class TestCase extends BaseTestCase
 {
     use CreatesApplication;
 
-    /*
-     * Signs in a user or creates a new one and signs it in.
-     *
-     * @param mixed $attributes User attributes or user model
-     */
-    protected function signIn($attributes = [])
-    {
-        $namespace = config('auth.providers.users.model');
-
-        if ($attributes instanceof $namespace) {
-            $this->actingAs($attributes);
-
-            return $this;
-        }
-
-        $this->actingAs(factory($namespace)->create($attributes));
-
-        return $this;
-    }
-
     protected function assertAdminOnly($method, $url)
     {
         $this->{$method}($url, [])
             ->assertRedirect(route('login'));
-        $this->signIn()->{$method}($url)
+        $this->be(User::factory()->create())->{$method}($url)
             ->assertRedirect('/');
     }
 }
